Discover movies, admin role + fixes & updates

+++ src/components/Users.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 import userService from '../services/users'
 import { Container, Typography } from '@mui/material'
 import { format, parseISO } from 'date-fns'
@@ -11,24 +11,54 @@ import { CancelOutlined, DeleteOutlined,
 import Redirect from './Redirect'
 import { useConfirm } from 'material-ui-confirm'
 import CustomToolbar from './CustomToolbar'
+import { useSnackbar } from 'notistack'
 
-const Users = () => {
+const Users = ({ user }) => {
   const [users, setUsers] = useState([])
   const [rowModesModel, setRowModesModel] = useState({})
   const [rowSelectionModel, setRowSelectionModel] = useState([])
   const [error, setError] = useState(null)
   const confirm = useConfirm()
+  const { enqueueSnackbar } = useSnackbar()
 
   useEffect(() => {
-    userService.getAllUsers()
-      .then(response => {
-        setUsers(response)
-      })
-      .catch(error => {
-        console.log(error)
-        setError(error)
-      })
+    if (user) {
+      if (user.role === 'moderator') {
+        userService.getUsers()
+          .then(response => {
+            setUsers(response)
+          })
+          .catch(error => {
+            console.log(error)
+            setError(error)
+          })
+      } else if (user.role === 'admin') {
+        userService.getAllUsers()
+          .then(response => {
+            setUsers(response)
+          })
+          .catch(error => {
+            console.log(error)
+            setError(error)
+          })
+      } else {
+        setError('Access denied')
+      }
+    } else {
+      setError('Access denied')
+    }
   }, [])
+
+  const columnVisibilityModel = useMemo(() => {
+    if (user.role === 'admin') {
+      return {
+        role: true
+      }
+    }
+    return {
+      role: false
+    }
+  }, [user])
 
   const handleRowEditStop = (params, event) => {
     if (params.reason === GridRowEditStopReasons.rowFocusOut) {
@@ -42,27 +72,32 @@ const Users = () => {
     setRowModesModel({ ...rowModesModel, [id]: { mode: GridRowModes.View } })
   }
   const handleDeleteClick = (id) => () => {
-    const user = users.find((u) => u.id === id)
-    confirm({
-      title: 'Delete user?',
-      description: user.name,
-      confirmationText: 'Delete', dialogProps: {
-        PaperProps: {
-          sx: {
-            width: 'auto',
-            minWidth: '13%'
+    if (user.role === 'admin') {
+      const user = users.find((u) => u.id === id)
+      confirm({
+        title: 'Delete user?',
+        description: user.name,
+        confirmationText: 'Delete', dialogProps: {
+          PaperProps: {
+            sx: {
+              width: 'auto',
+              minWidth: '13%'
+            }
           }
-        }
-      } })
-      .then(() => {
-        setUsers(users.filter((user) => user.id !== id))
-        userService.removeUser(id)
-          .then(() => {})
-          .catch((error) => {
-            console.log(error)
-          })
-      })
-      .catch(() => {})
+        } })
+        .then(() => {
+          setUsers(users.filter((user) => user.id !== id))
+          userService.removeUser(id)
+            .then(() => {})
+            .catch((error) => {
+              console.log(error)
+            })
+        })
+        .catch(() => {})
+    } else {
+      enqueueSnackbar('Only admins are allowed to delete users',
+        { variant: 'error' })
+    }
   }
   const handleCancelClick = (id) => () => {
     setRowModesModel({
@@ -76,11 +111,19 @@ const Users = () => {
     }
   }
   const processRowUpdate = (newRow) => {
-    userService.editUser(newRow.id, newRow)
-      .then(() => {})
-      .catch(error => {
-        console.log(error)
-      })
+    if (user.role === 'moderator') {
+      userService.editDisabled(newRow.id, newRow)
+        .then(() => {})
+        .catch(error => {
+          console.log(error)
+        })
+    } else if (user.role === 'admin') {
+      userService.editUser(newRow.id, newRow)
+        .then(() => {})
+        .catch(error => {
+          console.log(error)
+        })
+    }
     const updatedRow = { ...newRow, isNew: false }
     setUsers(users.map((row) => (row.id === newRow.id ? updatedRow : row)))
     return updatedRow
@@ -90,20 +133,21 @@ const Users = () => {
   }
 
   const columns = [
-    { type: 'number', field: 'id', headerName: 'ID', flex: 0.3, headerAlign: 'left', align: 'left' },
-    { field: 'name', headerName: 'Name', flex: 1, renderCell: (params) => (
+    { field: 'id', type: 'number', headerName: 'ID', flex: 0.3, headerAlign: 'left', align: 'left' },
+    { field: 'name', headerName: 'Name', flex: 0.8, renderCell: (params) => (
       <Link to={`/users/${params.row.id}`}>{params.value}</Link>
     ) },
-    { field: 'email', headerName: 'Email address', flex: 1.5 },
-    { field: 'disabled', headerName: 'Disabled?', flex: 0.5, editable: true,
-      type: 'boolean' },
-    { field: 'createdAt', headerName: 'Account created', flex: 1, valueFormatter: (params) => {
+    { field: 'email', headerName: 'Email address', flex: 1.2 },
+    { field: 'role', type: 'singleSelect', valueOptions: ['user', 'moderator'],
+      headerName: 'Role', flex: 0.5, editable: true },
+    { field: 'disabled', headerName: 'Disabled?', flex: 0.6, editable: true, type: 'boolean' },
+    { field: 'createdAt', headerName: 'Account created', flex: 0.7, valueFormatter: (params) => {
       if (params.value === null) {
         return ''
       }
       return format(parseISO(params.value), 'dd.MM.yyyy')
     } },
-    { field: 'reviews', headerName: 'Reviews', flex: 0.5 },
+    { field: 'reviews', headerName: 'Reviews', flex: 0.4 },
     { field: 'actions', type: 'actions', headerName: 'Actions', cellClassName: 'actions',
       getActions: ({ id }) => {
         const isInEditMode = rowModesModel[id]?.mode === GridRowModes.Edit
@@ -178,6 +222,7 @@ const Users = () => {
         columns={columns}
         editMode='row'
         rowModesModel={rowModesModel}
+        columnVisibilityModel={columnVisibilityModel}
         onRowModesModelChange={handleRowModesModelChange}
         onRowEditStop={handleRowEditStop}
         processRowUpdate={processRowUpdate}
